drummer card in game

+++ src/drummer.cpp
@@ -6,12 +6,8 @@
 
 DrummerCard::DrummerCard(){}
 
-void DrummerCard::useCard ( std::vector <Player> & players , int numPlayer, int number )
+void DrummerCard::useCard ( std::vector <Player> & players , int numPlayer, double number )
 {
     int sum = players[numPlayer].getScorePlayer();
-    players[numPlayer].setScorePlayer( sum * number );
+    players[numPlayer].setScorePlayer(sum * number);
 }
-// void DrummerCard::draw ( int x , int y )
-// {
-//     DrawTexture (myAsset.drummer  , x , y , BLACK);
-// }
